[fixed]カード別のダイアログ表示
