+++ src/components/Navigation.vue
@@ -122,7 +122,6 @@ export default {
       ['mdi-home', 'ホーム','/'],
       ['mdi-magnify', '検索','/search'],
       ['mdi-calendar', 'カレンダー','/calendar'],
-      ['mdi-calendar', 'カレンダー2','/calendar2'],
       ['mdi-bell', '通知','/notice'],
       ['mdi-account', 'プロフィール','/profile'],
       ['mdi-cog', '設定','/setting']
