+++ src/views/Calendar.vue
@@ -8,7 +8,7 @@
       <v-col  cols="8">
         <v-row align-content="space-between">
           <v-col cols="12" align="center">
-          <div class="display-4 " id="tex">{{value}}日</div>
+          <div class="display-4 ">{{picker}}日</div>
           </v-col>
           <v-col cols="12" align="center">
           <div class="display-3">体重:60kg</div>
@@ -18,20 +18,6 @@
           </v-col>
         </v-row>
       </v-col>
-      <!-- <v-col cols="12">
-        <v-row justify="space-around">
-          <v-switch v-model="landscape" class="ma-4" label="Landscape"></v-switch>
-          <v-switch v-model="reactive" class="ma-4" label="Reactive"></v-switch>
-          <v-switch v-model="flat" class="ma-4" label="Flat"></v-switch>
-          <v-switch v-model="fullWidth" class="ma-4" label="Full width"></v-switch>
-          <v-switch v-model="showCurrent" class="ma-4" label="Show current date"></v-switch>
-          <v-switch v-model="month" class="ma-4" label="Month picker"></v-switch>
-          <v-switch v-model="multiple" class="ma-4" label="Multiple"></v-switch>
-          <v-switch v-model="readonly" class="ma-4" label="Readonly"></v-switch>
-          <v-switch v-model="disabled" class="ma-4" label="Disabled"></v-switch>
-          <v-switch v-model="enableEvents" class="ma-4" label="Events"></v-switch>
-        </v-row>
-      </v-col> -->
       <v-col cols="4" align="right">
         <v-date-picker
           v-model="picker"
@@ -54,96 +40,96 @@
       <div class="display-2">今日のごはん</div>
     </v-row>
     <v-divider></v-divider>
-    <!--<button v-on:click="test">add</button>-->
+
     <v-row id="clmenu" align="center">
-    <!--<p id="clmenu"></p>-->
-          <v-col
-            v-for="card in cards"
-            :key="card.title"
-            :cols="card.flex"
-            @click.stop="card.dialog = true"
+      <v-col
+        v-for="(card,index) in todayMenu"
+        :key="index"
+        :cols=3
+        @click.stop="onClickBtn(card)"
+      >
+        <v-subheader>{{card.subtitle}}</v-subheader>
+        <v-card>
+          <v-img
+            :src="card.image"
+            class="white--text align-end"
+            gradient="to bottom, rgba(0,0,0,.1), rgba(0,0,0,.5)"
+            height="200px"
           >
-            <v-card v-if="card.dialog">
-              <v-subheader>{{card.ranking}}</v-subheader>
-              <v-img
-                :src="card.src"
-                class="white--text align-end"
-                gradient="to bottom, rgba(0,0,0,.1), rgba(0,0,0,.5)"
-                height="200px"
-              >
-                <v-card-title v-text="card.title"></v-card-title>
-              </v-img>
-              <v-card-actions>
-                <v-spacer></v-spacer>
-                <v-btn icon>
-                  <v-icon>mdi-fire</v-icon>
-                </v-btn>
-                <span>400kcal</span>
-              </v-card-actions>
-            </v-card>
-            <v-dialog
-              v-model="card.dialog"
-              max-width="600"
-            >
-              <v-card>
-              <v-img
-                :src= "card.src"
-                class="white--text align-end"
-              >
-              </v-img>
-              <v-card-title>{{card.title}}</v-card-title>
-              <v-divider class="mx-4"></v-divider>
-              <v-card-title>栄養成分</v-card-title>
-              <v-list class="transparent">
-                <v-list-item>
-                  <v-list-item-title>エネルギー</v-list-item-title>
-                  <v-list-item-subtitle>400kcal</v-list-item-subtitle>
-                </v-list-item>
-                <v-list-item>
-                  <v-list-item-title>炭水化物</v-list-item-title>
-                  <v-list-item-subtitle>400kcal</v-list-item-subtitle>
-                </v-list-item>
-                <v-list-item>
-                  <v-list-item-title>タンパク質</v-list-item-title>
-                  <v-list-item-subtitle>400kcal</v-list-item-subtitle>
-                </v-list-item>
-                <v-list-item>
-                  <v-list-item-title>脂質</v-list-item-title>
-                  <v-list-item-subtitle>400kcal</v-list-item-subtitle>
-                </v-list-item>
-                <v-list-item>
-                  <v-list-item-title>なんか</v-list-item-title>
-                  <v-list-item-subtitle>400kcal</v-list-item-subtitle>
-                </v-list-item>
-              </v-list>
-              <v-card-actions>
-                <v-spacer></v-spacer>
-                <v-btn icon>
-                  <v-icon>mdi-fire</v-icon>
-                </v-btn>
-                <span>400kcal</span>
-              </v-card-actions>
-            </v-card>
-            </v-dialog>
-          </v-col>
-        </v-row>
+            <v-card-title v-text="card.title"></v-card-title>
+          </v-img>
+
+          <v-card-actions>
+            <v-spacer></v-spacer>
+
+            <v-btn icon>
+              <v-icon>mdi-fire</v-icon>
+            </v-btn>
+            <span>{{card.calorele}}kcal</span>
+          </v-card-actions>
+        </v-card>
+      </v-col>
+
+      <v-dialog
+        v-model="dialog"
+        v-if="currentCard"
+        max-width="600"
+      >
+        <v-card>
+          <v-img
+            :src= "currentCard.image"
+            class="white--text align-end"
+          >
+          </v-img>
+          <v-card-title>{{currentCard.title}}</v-card-title>
+          <v-divider class="mx-4"></v-divider>
+          <v-card-title>栄養成分</v-card-title>
+          <v-list class="transparent">
+            <v-list-item>
+              <v-list-item-title>エネルギー</v-list-item-title>
+              <v-list-item-subtitle>400kcal</v-list-item-subtitle>
+            </v-list-item>
+            <v-list-item>
+              <v-list-item-title>炭水化物</v-list-item-title>
+              <v-list-item-subtitle>400kcal</v-list-item-subtitle>
+            </v-list-item>
+            <v-list-item>
+              <v-list-item-title>タンパク質</v-list-item-title>
+              <v-list-item-subtitle>400kcal</v-list-item-subtitle>
+            </v-list-item>
+            <v-list-item>
+              <v-list-item-title>脂質</v-list-item-title>
+              <v-list-item-subtitle>400kcal</v-list-item-subtitle>
+            </v-list-item>
+            <v-list-item>
+              <v-list-item-title>なんか</v-list-item-title>
+              <v-list-item-subtitle>400kcal</v-list-item-subtitle>
+            </v-list-item>
+          </v-list>
+
+          <v-card-actions>
+            <v-spacer></v-spacer>
+            <v-btn icon>
+              <v-icon>mdi-fire</v-icon>
+            </v-btn>
+            <span>{{currentCard.calorele}}cal</span>
+          </v-card-actions>
+        </v-card>
+      </v-dialog>  
+    </v-row>
+
   </div>
 </template>
+
 <script>
 import Navigation from '@/components/Navigation.vue';
 export default {
   components: {
     Navigation
   },
-  props: {
-    value: {
-      type: String,
-      default: new Date().toISOString().substr(0, 10)
-    }
-  },
    data () {
     return {
-      pickerr: new Date().toISOString().substr(0, 10),
+      picker: new Date().toISOString().substr(0, 10),
       landscape: false,
       reactive: true,
       fullWidth: true,
@@ -154,41 +140,45 @@ export default {
       readonly: false,
       disabled: false,
       enableEvents: true,
-      cards: [
-        { title: 'スパゲッティ', src: require('../assets/料理/料理.png'),dialog: false, flex: 3,ranking: '朝食',num:0,mdate:"2020-12-02"},
-        { title: 'ハンバーガー', src: require('../assets/料理/ハンバーガー.png'),dialog: false, flex: 3,ranking: '昼食',num:1,mdate:"2020-12-16"},
-        { title: '牛丼', src: require('../assets/料理/牛丼.png'),dialog: false, flex: 3,ranking: '夜食',num:2,mdate:"2020-12-02"},
-        { title: '牛丼', src: require('../assets/料理/牛丼.png'),dialog: false, flex: 3,ranking: '夜食',num:3,mdate:"2020-12-02"},
-        { title: '牛丼', src: require('../assets/料理/牛丼.png'),dialog: false, flex: 3,ranking: '夜食',num:4,mdate:"2020-12-02"},
-      ],
+
+      dialog: false,
+      currentCard: null,
+      foodMemos: [
+        { memoDate: "2020-12-02",title: "牛丼",image: require('../assets/料理/牛丼.png'),subtitle: "朝食",calorele: 100 },
+        { memoDate: "2020-12-02",title: "ハンバーガー",image: require('../assets/料理/ハンバーガー.png'),subtitle: "夜食",calorele: 200 },
+        { memoDate: "2020-12-06",title: "牛丼",image: require('../assets/料理/牛丼.png'),subtitle: "朝食",calorele: 500 },
+      ]
     }
   },
   computed: {
     functionEvents () {
       return this.month ? this.monthFunctionEvents : this.dateFunctionEvents
     },
-    picker: {
-      /*get() {
-        new Date().toISOString().substr(0, 10)
-        return this.value;
-      },*/
-      set(val) {
-        document.getElementById("tex").innerHTML=val+"日"
-        //alert(this.cards[0].mdate)
-        for(var cnt=0;cnt<this.cards.length;cnt++){
-          if(val===this.cards[cnt].mdate){
-            alert(val)
-          }
-        }
-        //this.$emit("input", val);
-      }
+    todayMenu(){
+      const menu = []
+      const today = this.picker
+      this.foodMemos.forEach(function (value) {
+        // console.log(index + '番目 : ' + value);
+        if(value.memoDate == today){menu.push(value)}
+      });
+      return menu
     }
   },
   methods: {
     dateFunctionEvents (date) {
+      // console.log("date:"+date)
+      // console.log("parseInt(day, 10):"+parseInt(day, 10))
       const [,, day] = date.split('-')
-      if ([12, 17, 28].includes(parseInt(day, 10))) return true
-      if ([1, 19, 22].includes(parseInt(day, 10))) return ['red', '#00f']
+      const memoDay = []
+      this.foodMemos.forEach(function (value) {
+        // console.log(index + '番目 : ' + value);
+        const [,, memo] = value.memoDate.split('-')
+        memoDay.push(Number(memo))
+      });
+      //[fix]ここら辺めっちゃ呼び出されてるけど、どうにかならないか？
+      console.log("memoDay:"+memoDay)
+      if (memoDay.includes(parseInt(day, 10))) return true
+      // if ([1, 19, 22].includes(parseInt(day, 10))) return ['red', '#00f']
       return false
     },
     monthFunctionEvents (date) {
@@ -197,6 +187,10 @@ export default {
       if ([2, 5, 12].includes(month)) return ['error', 'purple', 'rgba(0, 128, 0, 0.5)']
       return false
     },
+    onClickBtn(card) {
+      this.currentCard = card
+      this.dialog = true
+    }
   },
 };
 </script>
